Check for precedence between two activities.

If activity B occured, it should be preceded by activity A in the same case.

precedence(activity_a, activity_b)

Arguments

activity_a
Activity A. A character vector of length one. This should be an activity of the event log supplied to `check_rule`.

activity_b
Activity B. A character vector of length one. This should be an activity of the event log supplied to `check_rule`.

See also

Examples

library(bupaR)
library(eventdataR)

# A MRI Scan should be preceeded by a Blood test.

patients %>%
check_rule(precedence("Blood test","MRI SCAN"))

#> Log of 5442 events consisting of:
#> 7 traces 
#> 500 cases 
#> 2721 instances of 7 activities 
#> 7 resources 
#> Events occurred from 2017-01-02 11:41:53 until 2018-05-05 07:16:02 
#>  
#> Variables were mapped as follows:
#> Case identifier:		patient 
#> Activity identifier:		handling 
#> Resource identifier:		employee 
#> Activity instance identifier:	handling_id 
#> Timestamp:			time 
#> Lifecycle transition:		registration_type 
#> 
#> # A tibble: 5,442 x 8
#>    handling patient employee handling_id registration_ty~ time               
#>    <fct>    <chr>   <fct>    <chr>       <fct>            <dttm>             
#>  1 Registr~ 1       r1       1           start            2017-01-02 11:41:53
#>  2 Registr~ 2       r1       2           start            2017-01-02 11:41:53
#>  3 Registr~ 3       r1       3           start            2017-01-04 01:34:05
#>  4 Registr~ 4       r1       4           start            2017-01-04 01:34:04
#>  5 Registr~ 5       r1       5           start            2017-01-04 16:07:47
#>  6 Registr~ 6       r1       6           start            2017-01-04 16:07:47
#>  7 Registr~ 7       r1       7           start            2017-01-05 04:56:11
#>  8 Registr~ 8       r1       8           start            2017-01-05 04:56:11
#>  9 Registr~ 9       r1       9           start            2017-01-06 05:58:54
#> 10 Registr~ 10      r1       10          start            2017-01-06 05:58:54
#> # ... with 5,432 more rows, and 2 more variables: .order <int>,
#> #   precedence_Blood_test_MRI_SCAN <lgl>