Check succession between two activities

If activity A happens, it should be eventually followed by activity B. If activity B happens, it should be preceded by activity A.

succession(activity_a, activity_b)

Arguments

activity_a
Activity A. A character vector of length one. This should be an activity of the event log supplied to `check_rule`.

activity_b
Activity B. A character vector of length one. This should be an activity of the event log supplied to `check_rule`.

See also

Examples

library(bupaR)
library(eventdataR)

# Blood test should always happen before a MRI Scan,
# and both should happen when one of them happens.
patients %>%
check_rule(succession("Blood test","MRI SCAN"))

#> Event log consisting of:
#> 5442 events
#> 7 traces
#> 500 cases
#> 7 activities
#> 2721 activity instances
#> 
#> # A tibble: 5,442 x 8
#>    handling   patient employee handling_id registration_ty~ time               
#>    <fct>      <chr>   <fct>    <chr>       <fct>            <dttm>             
#>  1 Registrat~ 1       r1       1           start            2017-01-02 11:41:53
#>  2 Registrat~ 2       r1       2           start            2017-01-02 11:41:53
#>  3 Registrat~ 3       r1       3           start            2017-01-04 01:34:05
#>  4 Registrat~ 4       r1       4           start            2017-01-04 01:34:04
#>  5 Registrat~ 5       r1       5           start            2017-01-04 16:07:47
#>  6 Registrat~ 6       r1       6           start            2017-01-04 16:07:47
#>  7 Registrat~ 7       r1       7           start            2017-01-05 04:56:11
#>  8 Registrat~ 8       r1       8           start            2017-01-05 04:56:11
#>  9 Registrat~ 9       r1       9           start            2017-01-06 05:58:54
#> 10 Registrat~ 10      r1       10          start            2017-01-06 05:58:54
#> # ... with 5,432 more rows, and 2 more variables: .order <int>,
#> #   succession_Blood_test_MRI_SCAN <lgl>